IB/core: Add default GID for RoCE GID Cache
When RoCE is used, a default GID address should be generated
for every supported RoCE type. These default GID addresses are
generated based on the IPv6 link-local address, but in contrast
to the GID based on the regular IPv6 link-local (as we generate
GID per IP address), these GIDs are also available if the net
device is down (in order to support loopback).
Moreover, these default GID addresses can't be deleted.

Change-Id: I1742f095cf6eb5b9c605f5a9ae794ae154b470b3
Signed-off-by: Matan Barak <matanb@mellanox.com>