IB/core: Modify ib_verbs and cma in order to use roce_gid_table
Previously, we resolved the dmac and took the smac and vlan
from the resolved address. Changing that into finding a net
device that matches the IP and vlan of the network packet
and querying the RoCE GID table for this net device,
GID and GID type.

ocrdma driver changes were done by Somnath Kotur <Somnath.Kotur@Emulex.Com>

Signed-off-by: Matan Barak <matanb@mellanox.com>