IB/cma: Add configfs for rdma_cm
Users would like to control the behaviour of rdma_cm.
For example, old applications which doesn't set the
required RoCE gid type could be executed on RoCE V2
network types. In order to support this configuration,
we implement a configfs for cma.

In order to use the configfs, one needs to mount it and
mkdir <IB device name>. Users can't remove the IB device
module until they removed this directory.

The current support file is default_roce_mode which can
be either IB/RoCE V1 or RoCE V2.

Change-Id: I0fadaed6ec1e8f1c38f0c2236d782e383f2a628a
Signed-off-by: Matan Barak <matanb@mellanox.com>