IB/core: Support find sgid index using a filter function
Sometimes a sgid index need to be found based on variable parameters.
For example, when the CM gets a packet from network, it needs to
match a sgid_index that matches the appropriate L2 attributes
of a packet. Extending the cache's API to include Ethernet L2
attribute is problematic, since they may be vastly extended
in the future. As a result, we add a find function that
gets a user filter function and searches the GID table
until a match is found.

Change-Id: I63cd6e58a7c9e11880a5063e90291038f43a192e
Signed-off-by: Matan Barak <matanb@mellanox.com>