IB/mlx4: Lock with RCU instead of RTNL
The function eth_link_query_port() used to take the RTNL lock when
call to netdev_master_upper_dev_get() was necessary. This makes it
impossible to call this function with RTNL lock is held. Calling
netdev_master_upper_dev_get_rcu() and locking with RCU instead solve
this problem.

Change-Id: I1d62331625d445a5e8ed3fecf3baddb869e6d493
Signed-off-by: Moni Shoua <monis@mellanox.com>