net/ipv6: Export addrconf_ifid_eui48
For loopback purposes, RoCE devices should have a default GID in the
port GID table, even when the interface is down. In order to do so,
we use the IPv6 link local address which would have been genenrated
for the related Ethernet netdevice when it goes up as a default GID.

addrconf_ifid_eui48 is used to gernerate this address, export it.

Change-Id: I35634ce7787a13b993f0a903b2cf3e26155b7bd1
Signed-off-by: Matan Barak <matanb@mellanox.com>