net/bonding: Export bond_option_active_slave_get_rcu
Some consumers of the netdev events API would like to know who is the
active slave when a NETDEV_CHANGEUPPER or NETDEV_BONDING_FAILOVER
events occur. For example, when managing RoCE GIDs, GIDs based on the
bond's ips should only be set on the port which corresponds to active
slave netdevice.

Change-Id: Ia929de88e7dc13b930cb634c4bf9d6f22078e067
Signed-off-by: Matan Barak <matanb@mellanox.com>