IB/mlx5: Allow mapping the free running counter on PROT_EXEC
The current mlx5 code disallows mapping the free running counter of
mlx5 based hardwares when PROT_EXEC is set.
Although this behaviour is correct, Linux does add an implicit VM_EXEC
to the vm_flags if the READ_IMPLIES_EXEC bit is set in the process
personality. This happens for example if the process stack is
executable.
This causes libmlx5 to output a warning and prevents the user from
reading the free running clock.
A possible fix is to only allow PROT_EXEC if
current->personality & READ_IMPLIES_EXEC is set. However, we don't
think executing the init segment of the hardware is such a security
risk, so we skip this at all.

issue: 753886
Change-Id: I8f94267642eda6fc87f8e8d4ef782476f5b02f89
Signed-off-by: Matan Barak <matanb@mellanox.com>