IB/core: Use the new IDR and locking infrastructure in uverbs_cmd
The new infrastructure introduced a new locking and objects scheme.
We rewrite the current uverbs_cmd handlers to use them. The new
infrastructure needs a types definition in order to figure out how
to allocate/free resources, which is defined in upstream patches.

This patch refactores the following things:
(*) Instead of having a list per type, we use the ucontext's list
(*) The locking semantics are changed:
      Two commands might try to lock the same object. If this object
      is locked for exclusive object, any concurrent access will get
      -EBUSY. This makes the user serialize access.
(*) The completion channel FD is created by using the infrastructure.
    It's release function and release context are serialized by the
    infrastructure.
(*) The live flag is no longer required.
(*) User may need to assign the user_handle explicitly.

issue: 776663
Change-Id: Ia23b60d018766b248823ef1445fefd56132ad471
Signed-off-by: Matan Barak <matanb@mellanox.com>