IB/core: Add uverbs merge trees functionality
In order to have a more robust query system, we propose having
a parse tree that is unique for a device and represents only
the features this particular device support. This is done
by having a root specification tree per feature.
Before a device registers itself as an IB device, it merges
all these trees into one parsing tree. This parsing tree
is used to parse all user-space commands.
A user-space application could read this parse tree. This
tree represents which types, actions and attributes are
supported for this device.

This is based on the idea of
Jason Gunthorpe <jgunthorpe@obsidianresearch.com>

Signed-off-by: Matan Barak <matanb@mellanox.com>